Refactor Components to separate files
